--- app/src/main/java/com/twoday/todaytrip/adapter/SecondRecyclerViewAdapter.kt
@@ -1,4 +1,0 @@
-package com.twoday.todaytrip.adapter
-
-class SecondRecyclerViewAdapter {
-}
