--- app/src/main/java/com/twoday/todaytrip/adapter/ThirdRecyclerViewAdapter.kt
@@ -1,4 +1,0 @@
-package com.twoday.todaytrip.adapter
-
-class ThirdRecyclerViewAdapter {
-}
