+++ app/src/main/java/com/twoday/todaytrip/place_list_adapter/SecondRecyclerViewAdapter.kt
@@ -1,0 +1,4 @@
+package com.twoday.todaytrip.place_list_adapter
+
+class SecondRecyclerViewAdapter {
+}
